requirements-dev.txt
black
reorder-python-imports
flake8
mypy
types-requests
pre-commit
-r requirements.txt

pytest
requests-mock
pytest-cov